no op condenser in separate file as proof of concept
